Move timer macro out of root component into agent (review markup)
