+++ thoth-app/src/agent/version_timer.rs
@@ -1,0 +1,6 @@
+timer_agent! {
+    VersionTimerAgent,
+    VersionTimerDispatcher,
+    VersionTimerRequest,
+    VersionTimerResponse,
+}
